+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## (IN PROGRESS)
 
-## [2.1.0](https://github.com/folio-org/ui-export-manager/tree/v2.1.0) (2021-03-04)
+## [2.1.0](https://github.com/folio-org/ui-export-manager/tree/v2.1.0) (2022-03-04)
 [Full Changelog](https://github.com/folio-org/ui-export-manager/compare/v2.0.0...v2.1.0)
 
 * Search and filter organization exports logs. Refs UIEXPMGR-26.